chore: apply ktlint and detekt
