+++ app/src/main/java/com/lyfe/android/core/model/UserInfo.kt
@@ -1,7 +1,7 @@
 package com.lyfe.android.core.model
 
 data class UserInfo(
-    val id: Long,
-    val name: String,
-    val profileImage: String
+	val id: Long,
+	val name: String,
+	val profileImage: String
 )
