+++ app/src/main/java/com/lyfe/android/feature/profile/ProfileScreen.kt
@@ -3,7 +3,6 @@ package com.lyfe.android.feature.profile
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.aspectRatio
@@ -13,19 +12,11 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
-import androidx.compose.foundation.lazy.grid.rememberLazyGridState
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.PagerState
 import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.text.ClickableText
-import androidx.compose.material3.Divider
 import androidx.compose.material3.Tab
 import androidx.compose.material3.TabRow
 import androidx.compose.material3.TabRowDefaults
@@ -34,7 +25,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.key
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
@@ -52,7 +42,6 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
 import com.bumptech.glide.integration.compose.placeholder
@@ -60,7 +49,6 @@ import com.lyfe.android.R
 import com.lyfe.android.core.common.ui.component.LyfeButton
 import com.lyfe.android.core.common.ui.definition.LyfeButtonType
 import com.lyfe.android.core.common.ui.model.TabItem
-import com.lyfe.android.core.common.ui.theme.Grey100
 import com.lyfe.android.core.common.ui.theme.Grey200
 import com.lyfe.android.core.common.ui.theme.Grey300
 import com.lyfe.android.core.common.ui.theme.Grey500
@@ -70,7 +58,6 @@ import com.lyfe.android.core.common.ui.theme.pretenard
 import com.lyfe.android.core.model.UserInfo
 import com.lyfe.android.core.navigation.LyfeScreens
 import com.lyfe.android.core.navigation.navigator.LyfeNavigator
-import com.lyfe.android.feature.feed.FeedScreenCardView
 
 @Composable
 fun ProfileScreen(
@@ -138,11 +125,11 @@ private fun ProfileUserInfo(
 	navigator: LyfeNavigator,
 	userInfo: UserInfo
 ) {
-	Row (
+	Row(
 		modifier = Modifier
 			.height(48.dp)
 			.padding(horizontal = 20.dp)
-	){
+	) {
 		GlideImage(
 			modifier = Modifier
 				.fillMaxHeight()
@@ -165,7 +152,7 @@ private fun ProfileUserInfo(
 				fontSize = 20.sp,
 				fontWeight = FontWeight.W700,
 				lineHeight = 32.sp,
-				color = Color.Black,
+				color = Color.Black
 			)
 
 			if (userInfo.id <= 0) {
@@ -222,11 +209,13 @@ private fun ProfileUserPostTabContent(
 
 		if (isGuest) {
 			// 게스트는 로그인 유도창 띄우기
-			ProfileGuestLoginView(viewModel = viewModel, navigator = navigator)
+			ProfileGuestLoginView(
+				viewModel = viewModel,
+				navigator = navigator
+			)
 		}
 
 		ProfileUserPostPager(
-			navigator = navigator,
 			viewModel = viewModel,
 			isGuest = isGuest,
 			pagerState = pagerState
@@ -260,7 +249,7 @@ private fun ProfileTab(
 				text = {
 					Text(
 						text = tabItem.text,
-						style = getTabTextStyle(pagerState.currentPage, index),
+						style = getTabTextStyle(pagerState.currentPage, index)
 					)
 				},
 				selected = isCurrentTab(pagerState.currentPage, index),
@@ -279,7 +268,7 @@ private fun ProfileGuestLoginView(
 		modifier = Modifier.fillMaxSize(),
 		horizontalAlignment = CenterHorizontally
 	) {
-		Spacer(modifier = Modifier.fillMaxHeight(0.2f))
+		Spacer(modifier = Modifier.weight(1f))
 
 		Text(
 			text = stringResource(R.string.profile_screen_guest_login_title),
@@ -301,7 +290,7 @@ private fun ProfileGuestLoginView(
 			horizontalPadding = 24.dp
 		) {
 			// TODO 로그인 화면으로
-//			navigator.navigate(route = LyfeScreens.Login.route)
+// 			navigator.navigate(route = LyfeScreens.Login.route)
 			viewModel.updateToUserMode()
 		}
 
@@ -318,20 +307,20 @@ private fun ProfileGuestLoginView(
 			)
 		) {
 			// TODO 로그인 화면으로
-//			navigator.navigate(route = LyfeScreens.Login.route)
+			navigator.navigate(route = LyfeScreens.Login.route)
 		}
+
+		Spacer(modifier = Modifier.weight(2f))
 	}
 }
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
 private fun ProfileUserPostPager(
-	navigator: LyfeNavigator,
 	viewModel: ProfileViewModel,
 	isGuest: Boolean,
 	pagerState: PagerState
 ) {
-
 	HorizontalPager(
 		modifier = Modifier.fillMaxSize(),
 		state = pagerState
@@ -341,79 +330,15 @@ private fun ProfileUserPostPager(
 		when (page) {
 			0 -> {
 				// 신청 사진 리스트
-				ProfileUserImageFeedContent(
-					navigator = navigator,
+				ProfileImageFeedScreen(
 					viewModel = viewModel
 				)
 			}
 			1 -> {
 				// 고민 글
-				ProfileUserTextFeedContent(
-					navigator = navigator,
+				ProfileTextFeedScreen(
 					viewModel = viewModel
 				)
-			}
-		}
-	}
-}
-
-@Composable
-private fun ProfileUserImageFeedContent(
-	navigator: LyfeNavigator,
-	viewModel: ProfileViewModel,
-) {
-	val imageFeeds by viewModel.imageFeedList.collectAsStateWithLifecycle()
-	val lazyGridState = rememberLazyGridState()
-
-	LaunchedEffect(Unit) {
-		viewModel.fetchImageFeedList()
-	}
-
-	LazyVerticalGrid(
-		state = lazyGridState,
-		columns = GridCells.Fixed(2),
-		contentPadding = PaddingValues(horizontal = 20.dp, vertical = 16.dp),
-		horizontalArrangement = Arrangement.spacedBy(16.dp),
-		verticalArrangement = Arrangement.spacedBy(12.dp)
-	) {
-		items(imageFeeds) { imageFeed ->
-			key(imageFeed.feedId) {
-				FeedScreenCardView(
-					modifier = Modifier,
-					feed = imageFeed
-				)
-			}
-		}
-	}
-}
-
-@Composable
-private fun ProfileUserTextFeedContent(
-	navigator: LyfeNavigator,
-	viewModel: ProfileViewModel
-) {
-	val textFeeds by viewModel.textFeedList.collectAsStateWithLifecycle()
-	val lazyListState = rememberLazyListState()
-
-	LaunchedEffect(Unit) {
-		viewModel.fetchTextFeedList()
-	}
-
-	LazyColumn(
-		state = lazyListState,
-		contentPadding = PaddingValues(horizontal = 20.dp, vertical = 16.dp),
-		verticalArrangement = Arrangement.spacedBy(12.dp)
-	) {
-		items(textFeeds) { textFeed ->
-			key(textFeed.feedId) {
-				ProfileScreenTextFeedView(
-					modifier = Modifier,
-					feed = textFeed
-				)
-
-				Spacer(modifier = Modifier.height(12.dp))
-
-				Divider(color = Grey100, thickness = 1.dp)
 			}
 		}
 	}
