+++ app/src/main/java/com/lyfe/android/feature/setting/Setting.kt
@@ -1,9 +1,9 @@
 package com.lyfe.android.feature.setting
 
 enum class Setting {
-    NOTIFICATION,
-    SCRAP,
-    USER_EXPERIENCE,
-    PRIVACY_POLICY,
-    DELETE_ACCOUNT
+	NOTIFICATION,
+	SCRAP,
+	USER_EXPERIENCE,
+	PRIVACY_POLICY,
+	DELETE_ACCOUNT
 }
